app/requirements.txt
requests
django
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
isort
flake8
flake8-quotes
gunicorn
psycopg2-binary
django-environ